make clang happy. Again

+++ internal/libraw_x3f.cpp
@@ -1401,7 +1401,7 @@ static void huffman_decode_row(x3f_info_t *I,
   x3f_image_data_t *ID = &DEH->data_subsection.image_data;
   x3f_huffman_t *HUF = ID->huffman;
 
-  int16_t c[3] = {offset,offset,offset};
+  int16_t c[3] = {(int16_t)offset,(int16_t)offset,(int16_t)offset};
   int col;
   bit_state_t BS;
